Create README.md
This repo is an implementation of the Image Augmentation Loop with BNN (Binary Neural Net) on the CIFAR-10 dataset, using PyTorch.
Using Google Colab, you can execute the notebook using:

%run Image_Augmentation_BNN/main_complete.py --model vgg_cifar10_binary --save VGG16 --dataset cifar10 --epochs 55 --gpu 0 --lr 1

Of course, you can modify the flag values or add other flags to ajust the parameters. The flags description is in the main.complete.py file.


By Simon Beaulac.